% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalcusp.R
\name{chi2pvalue}
\alias{chi2pvalue}
\title{Chi square test of linear vs CUSP model}
\usage{
chi2pvalue(fitcusp)
}
\arguments{
\item{fitcusp}{output of the cusp() function}
}
\value{
the chi-square values of the likelihood-ratio test
}
\description{
This function compute the p-value comparing
the linear vs cusp model using Chi-square test (likelihood-ratio test)
}
\examples{
data(ecosystem_ns)
fitc <- cusp(y ~ PC1, alpha ~ Fishing,
            beta ~ Temperature,  data=ecosystem_ns)
chi2pvalue(fitc)

}
