% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_nea}
\alias{cod_nea}
\title{Biomass of North East arctic cod}
\format{
A data frame with 71 rows (corresponding to 71 years) and 4 variables:
\describe{
  \item{Year}{Year of the measurement}
  \item{Biomass}{Spawning stock biomass data deriving from stock assessment}
  \item{Fishing}{Fishing mortality estimated by stock assessment}
  \item{Temperature}{Temperature from NOAA ErSST v5 in degree celsius}
}
}
\source{
Sguotti, C. et al. 2019
}
\usage{
cod_nea
}
\description{
A dataset containing time series of the spawning stock biomass
of the North-East Artic Sea cod stock between 1946 and 2016
}
\references{
Sguotti, C. et al. Catastrophic dynamics limit Atlantic cod recovery. Proceedings of the Royal Society B 286, (2019).
}
\keyword{datasets}
