% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codnea}
\alias{codnea}
\title{Biomass of North East arctic cod}
\format{
A data frame with 71 rows (corresponding to 71 years) and 4 variables:
\describe{
  \item{Year}{Year of the measurment}
  \item{Biomass}{Spawning stock biomass, in tonnes?}
  \item{Fishing}{Fishing pressure estimated by ...}
  \item{Temperature}{Temperature from ... in degree celsius}
}
}
\source{
ICESstock assessment?? \url{https://www.ices.dk/}
}
\usage{
codnea
}
\description{
A dataset containing time series of the spawning stock biomass
of the NorthEast artic cod stock between 1946 and 2016
}
\references{
If any
}
\keyword{datasets}
