% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecosystem_ns}
\alias{ecosystem_ns}
\title{Ecosystem state of the North Sea}
\format{
A data frame with 29 rows (corresponding to 29 years) and 4 variables:
\describe{
  \item{Year}{Year of the measurment}
  \item{PC1}{Ecosystem state from a PCA on fish and plankton biomass}
  \item{Fishing}{Fishing pressure estimated by Couce et al. 2019, in hours}
  \item{Temperature}{Temperature from NOAA ErSST v5 in degree celsius}
}
}
\source{
Sguotti et al., 2022
}
\usage{
ecosystem_ns
}
\description{
A dataset containing time series of the North Sea community state index from 1985 to 2019.
It was built performing a PCA on a matrix of data coming from the
Continuous Plankton Recorder (CPR) and International Bottom Trawl Survey (IBTS)
The Temperature is the sea surface temperature (expressed in degree C)
derived from NOAA ErSST v5 and fishing effort (expressed as hours swept per year)
from Couce at al. 2019.
}
\references{
Couce et al. 2019
}
\keyword{datasets}
