% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalcusp.R
\name{evalcusp}
\alias{evalcusp}
\title{Evaluation of the CUSP model vs linear model}
\usage{
evalcusp(fitcusp)
}
\arguments{
\item{fitcusp}{output of the cusp() function}
}
\value{
A data.frame with
\describe{
  \item{rsquared}{the pseudo R2 of the cusp model (should be >0.3)}
  \item{delta.aicc}{the minimum difference between AICc of the linear model and cusp model (should be >0)}
  \item{percin}{percentage of points within the cusp area (should be >10)}
  \item{pval.state}{p-value of the state variable (should be <0.05)}
}
}
\description{
This function automatically return values to
pre-validate the cusp model
}
\examples{
data(ecosystem_ns)
fitc <- cusp(y ~ PC1, alpha ~ Fishing,
            beta ~ Temperature,  data=ecosystem_ns)
evalcusp(fitc)

}
