% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishtraits_med}
\alias{fishtraits_med}
\title{Traits of landings of Mediterranean Sea fish}
\format{
A data frame with 71 rows (corresponding to 71 years) and 4 variables:
\describe{
  \item{Year}{Year of the measurement}
  \item{PC1}{PC1 of the community weighted mean traits of the landings}
  \item{Fishing}{Fishing capacity estimated for the entire Mediterranean Sea by FAO}
  \item{Temperature}{Sea surface temperature from the NASA PODAAC}
}
}
\source{
FAO Mediterranean fisheries landings \url{http://www.fao.org/fishery/statistics/GFCM-capture-production/en}
}
\usage{
fishtraits_med
}
\description{
This dataset was derived by combining the Mediterranean fisheries landings data
of the FAO for the years 1985–2015 with a matrix including data on 23 traits
related to the biology, ecology, trophic role, distribution, habitat and behaviour
of 205 species (mostly fish, but also mollusks and crustaceans).
The multiplication of the two datasets resulted into a matrix of
community weighted mean traits landings by year in the Mediterranean Sea that was analysed in a PCA.
}
\references{
Tsimara, E. et al. An Integrated Traits Resilience Assessment of Mediterranean fisheries landings. Journal of Animal Ecology 90, 2122–2134 (2021).

Koutsidi, M. et al. (2019). 23 biological traits of 235 species. figshare. Dataset. https://doi.org/10.6084/m9.figshare.11347406.v1
}
\keyword{datasets}
