% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRA.R
\name{plotRA}
\alias{plotRA}
\title{Plot Resilience landscape}
\usage{
plotRA(
  var,
  fitcusp,
  pal = rev(grDevices::terrain.colors(11))[-1],
  main = "",
  ...
)
}
\arguments{
\item{var}{output of the cusp() function}

\item{fitcusp}{output of the cusp() function}

\item{pal}{output of the cusp() function}

\item{main}{title of the plot}

\item{...}{other arguments sent to plotCuspBifurcation}
}
\description{
This function plot the resilience landscape
based on the output of the cuspra function
and the plotting function from cusp package.
}
\examples{
data(codnea)
fitc <- cusp(y ~ Biomass, alpha ~ Fishing,
            beta ~ Temperature,  data=codnea)
ra <- cuspra(fitc)
plotRA(ra$cuspRA, fitc)

}
\keyword{assessment}
\keyword{resilience}
