% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuspra.R
\name{cuspra}
\alias{cuspra}
\title{Resilience Assessment of the CUSP output}
\usage{
cuspra(fitcusp)
}
\arguments{
\item{fitcusp}{output of the cusp() function}
}
\description{
This function compute the resilience assessment based
on the output of cusp function
}
\examples{
data(codnea)
fitc <- cusp(y ~ Biomass, alpha ~ Fishing,
            beta ~ Temperature,  data=codnea)
ra <- cuspra(fitc)

}
\keyword{assessment}
\keyword{resilience}
