% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotra.R
\name{plotra}
\alias{plotra}
\title{Plot Resilience landscape}
\usage{
plotra(
  var,
  fitcusp,
  pal = rev(grDevices::terrain.colors(11))[-1],
  main = "",
  ...
)
}
\arguments{
\item{var}{output of the cusp() function}

\item{fitcusp}{output of the cusp() function}

\item{pal}{output of the cusp() function}

\item{main}{title of the plot}

\item{...}{other arguments sent to plotCuspBifurcation}
}
\description{
This function plot the resilience landscape
based on the output of the cuspra function
and the plotting function from cusp package.
}
\examples{
data(ecosystem_ns)
fitc <- cusp(y ~ PC1, alpha ~ Fishing,
            beta ~ Temperature,  data=ecosystem_ns)
ra <- cuspra(fitc)
plotra(ra$cuspRA, fitc)

}
\keyword{assessment}
\keyword{resilience}
